@@ -1,47 +1,46 @@
 {
   "name": "medicapp-react",
-  "version": "1.3.0",
+  "version": "1.5.1",
   "private": true,
   "dependencies": {
-    "@ant-design/compatible": "1.0.8",
-    "@ant-design/icons": "4.6.2",
-    "@fullcalendar/daygrid": "5.8.0",
-    "@fullcalendar/react": "5.8.0",
-    "@south-paw/react-vector-maps": "3.1.0",
-    "@types/jest": "26.0.24",
-    "@types/node": "16.3.1",
-    "@types/react": "17.0.14",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.1.8",
-    "antd": "4.16.7",
-    "axios": "0.21.1",
-    "chart.js": "2.9.4",
+    "@ant-design/compatible": "^5.1.1",
+    "@ant-design/icons": "^5.1.3",
+    "@fullcalendar/daygrid": "6.1.7",
+    "@fullcalendar/react": "6.1.7",
+    "@south-paw/react-vector-maps": "3.2.0",
+    "@types/jest": "29.5.1",
+    "@types/node": "20.1.1",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
+    "@types/react-router-dom": "5.3.3",
+    "antd": "^5.5.1",
+    "axios": "1.4.0",
+    "chart.js": "4.3.0",
     "d3-scale-chromatic": "3.0.0",
-    "echarts": "5.1.2",
-    "echarts-for-react": "3.0.1",
+    "echarts": "^5.4.2",
+    "echarts-for-react": "3.0.2",
     "formik": "2.2.9",
-    "framer-motion": "4.1.17",
-    "google-map-react": "2.1.10",
-    "history": "5.0.0",
-    "node-sass": "6.0.1",
-    "react": "17.0.2",
-    "react-chartjs-2": "2.11.1",
+    "framer-motion": "10.12.8",
+    "google-map-react": "2.2.0",
+    "history": "5.3.0",
+    "node-sass": "^8.0.0",
+    "react": "18.2.0",
+    "react-chartjs-2": "5.2.0",
     "react-color": "2.19.3",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.4",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.3",
-    "recharts": "2.0.10",
-    "redux": "4.1.0",
-    "redux-thunk": "2.3.0",
-    "typescript": "4.3.5",
-    "yup": "0.32.9"
+    "react-dom": "18.2.0",
+    "react-redux": "8.0.5",
+    "react-router-dom": "6.11.1",
+    "react-scripts": "^5.0.1",
+    "recharts": "2.6.0",
+    "redux": "4.2.1",
+    "redux-thunk": "2.4.2",
+    "typescript": "4.9.5",
+    "yup": "^1.1.1"
   },
   "scripts": {
-    "start": "react-app-rewired start --jsx react",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test --env=jsdom",
-    "eject": "react-scripts eject"
+    "start": "craco start --jsx react",
+    "build": "craco build",
+    "test": "craco test --env=jsdom"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -67,14 +66,17 @@
     "trailingComma": "none"
   },
   "devDependencies": {
-    "@types/history": "4.7.9",
-    "@types/react-redux": "7.1.18",
-    "@types/yup": "0.29.13",
-    "babel-plugin-import": "1.13.3",
-    "customize-cra": "1.0.0",
-    "prettier": "2.3.2",
-    "react-app-rewire-less-modules": "1.3.0",
-    "react-app-rewired": "2.1.8",
-    "redux-devtools": "3.5.0"
+    "@craco/craco": "^7.1.0",
+    "@types/history": "4.7.11",
+    "@types/react-redux": "7.1.25",
+    "@types/yup": "0.29.14",
+    "craco-less": "^3.0.1",
+    "css-loader": "^6.7.3",
+    "less": "^4.1.3",
+    "less-loader": "^11.1.0",
+    "mini-css-extract-plugin": "^2.7.5",
+    "postcss-loader": "^7.3.0",
+    "prettier": "2.8.8",
+    "style-loader": "^3.3.2"
   }
 }
