@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
 import { store } from './redux/store';
@@ -9,15 +8,13 @@ import App from './App';
 
 import * as serviceWorker from './serviceWorker';
 import { HashRouter } from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
-  <Provider store={store}>
-    <HashRouter>
-      <App />
-    </HashRouter>
-  </Provider>,
-  document.getElementById('root')
-);
+createRoot(document.getElementById('root')).render( <Provider store={store}>
+  <HashRouter>
+    <App />
+  </HashRouter>
+</Provider>)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
