@@ -1,15 +1,15 @@
 export interface IPatient {
+  name: string;
+  age: number;
+  address: string;
+  number: string;
   id?: string;
   img?: string | ArrayBuffer;
   profileImg?: string | ArrayBuffer;
-  name: string;
   lastName?: string;
   fullName?: string;
-  number: string;
   phone?: string;
-  age: number;
   gender: string;
-  address: string;
   status?: string;
   lastVisit?: string;
 }
@@ -19,8 +19,7 @@ export interface IAppointment {
   name: string;
   email: string;
   number: string;
-  date: string;
-  fromTo: string;
+  date: string | null;
   doctor: string;
   injury: string;
   from: string;
