@@ -4,15 +4,13 @@ const piePatternImg = new Image();
 piePatternImg.src = piePatternSrc;
 
 const pieStyle = {
-  normal: {
-    opacity: 0.7,
-    color: {
-      image: piePatternImg,
-      repeat: 'repeat'
-    },
-    borderWidth: 3,
-    borderColor: '#336cfb'
-  }
+  opacity: 0.7,
+  color: {
+    image: piePatternImg,
+    repeat: 'repeat'
+  },
+  borderWidth: 3,
+  borderColor: '#336cfb'
 };
 
 const patientsAgeOptions = {
@@ -35,20 +33,14 @@ const patientsAgeOptions = {
       clockwise: true,
       radius: [60, '90%'],
       label: {
-        normal: {
-          position: 'inner',
-          textStyle: {
-            fontSize: 14,
-            fontWeight: 700,
-            color: '#000'
-          }
-        }
+        position: 'inner',
+        fontSize: 14,
+        fontWeight: 700,
+        color: '#000'
       },
       labelLine: {
-        normal: {
-          lineStyle: {
-            color: '#336cfb'
-          }
+        lineStyle: {
+          color: '#336cfb'
         }
       },
       data: [
@@ -83,20 +75,14 @@ const patientsGenderOptions = {
       clockwise: true,
       radius: [0, '90%'],
       label: {
-        normal: {
-          position: 'inner',
-          textStyle: {
-            fontSize: 14,
-            fontWeight: 700,
-            color: '#000'
-          }
-        }
+        position: 'inner',
+        fontSize: 14,
+        fontWeight: 700,
+        color: '#000'
       },
       labelLine: {
-        normal: {
-          lineStyle: {
-            color: '#336cfb'
-          }
+        lineStyle: {
+          color: '#336cfb'
         }
       },
       data: [
@@ -126,9 +112,7 @@ const departmentsOptions = {
       radius: [0, '90%'],
       roseType: 'area',
       label: {
-        normal: {
-          show: false
-        }
+        show: false
       },
       data: [
         { value: 115, name: 'Cardiology' },
