@@ -54,7 +54,7 @@ const DoctorsPage = () => {
       <PageAction onClick={openModal} icon='icofont-contact-add' />
 
       <Modal
-        visible={addingDoctor}
+        open={addingDoctor}
         onCancel={closeModal}
         title={<h3 className='title'>Add doctor</h3>}
         destroyOnClose
