@@ -1,36 +1,154 @@
 import React from 'react';
 
 import { Card } from 'antd';
-import { Icon } from '@ant-design/compatible';
+import
+  Icon,
+{
+  BellOutlined,
+  BookOutlined,
+  BugOutlined,
+  BulbOutlined,
+  BuildOutlined,
+  CalculatorOutlined,
+  CalendarOutlined,
+  CameraOutlined,
+  CarOutlined,
+  CarryOutOutlined,
+  CloudOutlined,
+  CodeOutlined,
+  DislikeOutlined,
+  EnvironmentOutlined,
+  ExperimentOutlined,
+  EyeOutlined,
+  FileOutlined,
+  FilterOutlined,
+  FireOutlined,
+  FlagOutlined,
+  FolderOutlined,
+  FolderOpenOutlined,
+  BellFilled,
+  BookFilled,
+  BugFilled,
+  BulbFilled,
+  BuildFilled,
+  CalculatorFilled,
+  CalendarFilled,
+  CameraFilled,
+  CarFilled,
+  CarryOutFilled,
+  CloudFilled,
+  CodeFilled,
+  DislikeFilled,
+  EnvironmentFilled,
+  ExperimentFilled,
+  EyeFilled,
+  FileFilled,
+  FilterFilled,
+  FireFilled,
+  FlagFilled,
+  FolderFilled,
+  FolderOpenFilled,
+  BellTwoTone,
+  BookTwoTone,
+  BugTwoTone,
+  BulbTwoTone,
+  BuildTwoTone,
+  CalculatorTwoTone,
+  CalendarTwoTone,
+  CameraTwoTone,
+  CarTwoTone,
+  CarryOutTwoTone,
+  CloudTwoTone,
+  CodeTwoTone,
+  DislikeTwoTone,
+  EnvironmentTwoTone,
+  ExperimentTwoTone,
+  EyeTwoTone,
+  FileTwoTone,
+  FilterTwoTone,
+  FireTwoTone,
+  FlagTwoTone,
+  FolderTwoTone,
+  FolderOpenTwoTone,
+}
+  from '@ant-design/icons';
 
 import { IPageData } from '../../interfaces/page';
 import { usePageData } from '../../hooks/usePage';
 
-const colors = ['#48bb78', '#336cfb', '#fc8b37', '#ed253d', '#ed64a6'];
+const outlinedIcons = [
+  BellOutlined,
+  BookOutlined,
+  BugOutlined,
+  BulbOutlined,
+  BuildOutlined,
+  CalculatorOutlined,
+  CalendarOutlined,
+  CameraOutlined,
+  CarOutlined,
+  CarryOutOutlined,
+  CloudOutlined,
+  CodeOutlined,
+  DislikeOutlined,
+  EnvironmentOutlined,
+  ExperimentOutlined,
+  EyeOutlined,
+  FileOutlined,
+  FilterOutlined,
+  FireOutlined,
+  FlagOutlined,
+  FolderOutlined,
+  FolderOpenOutlined,
+];
 
-const iconTypes = [
-  'bell',
-  'book',
-  'bug',
-  'bulb',
-  'calculator',
-  'build',
-  'calendar',
-  'camera',
-  'car',
-  'carry-out',
-  'cloud',
-  'code',
-  'dislike',
-  'environment',
-  'experiment',
-  'eye',
-  'file',
-  'filter',
-  'fire',
-  'flag',
-  'folder',
-  'folder-open'
+const filledIcons = [
+  BellFilled,
+  BookFilled,
+  BugFilled,
+  BulbFilled,
+  BuildFilled,
+  CalculatorFilled,
+  CalendarFilled,
+  CameraFilled,
+  CarFilled,
+  CarryOutFilled,
+  CloudFilled,
+  CodeFilled,
+  DislikeFilled,
+  EnvironmentFilled,
+  ExperimentFilled,
+  EyeFilled,
+  FileFilled,
+  FilterFilled,
+  FireFilled,
+  FlagFilled,
+  FolderFilled,
+  FolderOpenFilled,
+];
+
+const twoToneIcons = [
+  BellTwoTone,
+  BookTwoTone,
+  BugTwoTone,
+  BulbTwoTone,
+  BuildTwoTone,
+  CalculatorTwoTone,
+  CalendarTwoTone,
+  CameraTwoTone,
+  CarTwoTone,
+  CarryOutTwoTone,
+  CloudTwoTone,
+  CodeTwoTone,
+  DislikeTwoTone,
+  EnvironmentTwoTone,
+  ExperimentTwoTone,
+  EyeTwoTone,
+  FileTwoTone,
+  FilterTwoTone,
+  FireTwoTone,
+  FlagTwoTone,
+  FolderTwoTone,
+  FolderOpenTwoTone,
 ];
 
 const pageData: IPageData = {
@@ -56,8 +174,8 @@ const IconsOptionsPage = () => {
 
   const displayIcons = (icons, props) => (
     <div className='elem-list'>
-      {icons.map((icon) => (
-        <Icon key={icon} type={icon} {...props} />
+      {icons.map((icon, index) => (
+        <Icon key={index} component={icon} {...props} />
       ))}
     </div>
   );
@@ -66,47 +184,32 @@ const IconsOptionsPage = () => {
   const mediumFZ = { fontSize: 20 };
   const largeFZ = { fontSize: 24 };
 
-  const coloredIcons = (icons, props, colors) => (
-    <div className='elem-list'>
-      {icons.map((icon, index) => (
-        <Icon
-          key={icon}
-          type={icon}
-          theme='twoTone'
-          twoToneColor={colors[index % colors.length]}
-          {...props}
-        />
-      ))}
-    </div>
-  );
-
   return (
     <>
-      <Card title='Default icons'>{displayIcons(iconTypes, null)}</Card>
+      <Card title='Default icons'>
+        {displayIcons(outlinedIcons, null)}
+      </Card>
 
       <Card title='Icon sizes'>
         <h6 className='section-title mt-0'>16px font size</h6>
-        {displayIcons(iconTypes, { style: smallFZ })}
+        {displayIcons(outlinedIcons, { style: smallFZ })}
 
         <h6 className='section-title'>20px font size</h6>
-        {displayIcons(iconTypes, { style: mediumFZ })}
+        {displayIcons(outlinedIcons, { style: mediumFZ })}
 
         <h6 className='section-title'>24px font size</h6>
-        {displayIcons(iconTypes, { style: largeFZ })}
+        {displayIcons(outlinedIcons, { style: largeFZ })}
       </Card>
 
       <Card title='Icon theme' className='mb-0'>
         <h6 className='section-title mt-0'>Outlined</h6>
-        {displayIcons(iconTypes, { style: mediumFZ })}
+        {displayIcons(outlinedIcons, { style: mediumFZ })}
 
         <h6 className='section-title'>Filled</h6>
-        {displayIcons(iconTypes, { style: mediumFZ, theme: 'filled' })}
+        {displayIcons(filledIcons, { style: mediumFZ })}
 
         <h6 className='section-title'>Two tones</h6>
-        {displayIcons(iconTypes, { style: mediumFZ, theme: 'twoTone' })}
-
-        <h6 className='section-title'>Two tones colored</h6>
-        {coloredIcons(iconTypes, { style: mediumFZ }, colors)}
+        {displayIcons(twoToneIcons, { style: mediumFZ })}
       </Card>
     </>
   );
