@@ -1,13 +1,9 @@
 import React from 'react';
-
-import { Button, Form, Input, Switch } from 'antd';
-
-import PublicLayout from '../../layout/public/Public';
 import { Link } from 'react-router-dom';
+import { Button, Form, Input, Switch } from 'antd';
 import { useForm } from 'antd/es/form/Form';
+import PublicLayout from '../../layout/public/Public';
 import { useNavigateHome } from '../../utils/use-navigate-home';
-
-const { Item } = Form;
 
 const SignUp = () => {
   const navigateHome = useNavigateHome();
@@ -26,11 +22,11 @@ const SignUp = () => {
       <p className='text-color-200'>Create your Account</p>
 
       <Form form={form} layout='vertical' className='mb-5'>
-        <Item name='name' rules={[{ required: true, message: <></> }]}>
+        <Form.Item name='name' rules={[{ required: true, message: <></> }]}>
           <Input placeholder='Name' />
-        </Item>
+        </Form.Item>
 
-        <Item
+        <Form.Item
           name='email'
           rules={[
             { required: true, message: <></> },
@@ -38,11 +34,11 @@ const SignUp = () => {
           ]}
         >
           <Input placeholder='Email address' type='mail' />
-        </Item>
+        </Form.Item>
 
-        <Item name='password' rules={[{ required: true, message: <></> }]}>
+        <Form.Item name='password' rules={[{ required: true, message: <></> }]}>
           <Input placeholder='Password' type='password' />
-        </Item>
+        </Form.Item>
 
         <div className='d-flex align-items-center mb-4'>
           <Switch defaultChecked /> <span className='ml-2'>I agree to the Terms and Privacy.</span>
@@ -58,7 +54,7 @@ const SignUp = () => {
       </Form>
 
       <p>
-        Have an account? <Link to='sign-in'>Sign in!</Link>
+        Have an account? <Link to='../sign-in'>Sign in!</Link>
       </p>
     </PublicLayout>
   );
