@@ -1,28 +1,38 @@
 import React from 'react';
-import { Table, Badge, Menu, Dropdown } from 'antd';
+import { Table, Badge, Dropdown, TableColumnsType, Space } from 'antd';
 import { DownOutlined } from '@ant-design/icons/lib';
 
-const menu = (
-  <Menu>
-    <Menu.Item>Action 1</Menu.Item>
-    <Menu.Item>Action 2</Menu.Item>
-  </Menu>
-);
+interface DataType {
+  key: React.Key;
+  name: string;
+  platform: string;
+  version: string;
+  upgradeNum: number;
+  creator: string;
+  createdAt: string;
+}
+
+interface ExpandedDataType {
+  key: React.Key;
+  date: string;
+  name: string;
+  upgradeNum: string;
+}
+
+const items = [
+  { key: '1', label: 'Action 1' },
+  { key: '2', label: 'Action 2' },
+];
 
 const NestedTable = () => {
   const expandedRowRender = () => {
-    const columns = [
+    const columns: TableColumnsType<ExpandedDataType> = [
       { title: 'Date', dataIndex: 'date', key: 'date' },
       { title: 'Name', dataIndex: 'name', key: 'name' },
       {
         title: 'Status',
         key: 'state',
-        render: () => (
-          <span>
-            <Badge status='success' />
-            Finished
-          </span>
-        )
+        render: () => <Badge status="success" text="Finished" />,
       },
       { title: 'Upgrade Status', dataIndex: 'upgradeNum', key: 'upgradeNum' },
       {
@@ -30,57 +40,51 @@ const NestedTable = () => {
         dataIndex: 'operation',
         key: 'operation',
         render: () => (
-          <span className='table-operation'>
-            <Dropdown overlay={menu}>
-              <a onClick={(e) => e.preventDefault()} href='#'>
+          <Space size="middle">
+            <a>Pause</a>
+            <a>Stop</a>
+            <Dropdown menu={{ items }}>
+              <a>
                 More <DownOutlined />
               </a>
             </Dropdown>
-          </span>
-        )
-      }
+          </Space>
+        ),
+      },
     ];
 
     const data = [];
     for (let i = 0; i < 3; ++i) {
       data.push({
-        key: i,
+        key: i.toString(),
         date: '2014-12-24 23:12:00',
         name: 'This is production name',
-        upgradeNum: 'Upgraded: 56'
+        upgradeNum: 'Upgraded: 56',
       });
     }
     return <Table columns={columns} dataSource={data} pagination={false} />;
   };
 
-  const columns = [
+  const columns: TableColumnsType<DataType> = [
     { title: 'Name', dataIndex: 'name', key: 'name' },
     { title: 'Platform', dataIndex: 'platform', key: 'platform' },
     { title: 'Version', dataIndex: 'version', key: 'version' },
     { title: 'Upgraded', dataIndex: 'upgradeNum', key: 'upgradeNum' },
     { title: 'Creator', dataIndex: 'creator', key: 'creator' },
     { title: 'Date', dataIndex: 'createdAt', key: 'createdAt' },
-    {
-      title: 'Action',
-      key: 'operation',
-      render: () => (
-        <a onClick={(e) => e.preventDefault()} href='#'>
-          Publish
-        </a>
-      )
-    }
+    { title: 'Action', key: 'operation', render: () => <a>Publish</a> },
   ];
 
-  const data = [];
+  const data: DataType[] = [];
   for (let i = 0; i < 3; ++i) {
     data.push({
-      key: i,
-      name: 'Screem',
+      key: i.toString(),
+      name: 'Screen',
       platform: 'iOS',
       version: '10.3.4.5654',
       upgradeNum: 500,
       creator: 'Jack',
-      createdAt: '2014-12-24 23:12:00'
+      createdAt: '2014-12-24 23:12:00',
     });
   }
 
@@ -89,7 +93,7 @@ const NestedTable = () => {
       pagination={false}
       className='components-table-demo-nested'
       columns={columns}
-      expandedRowRender={expandedRowRender}
+      expandable={{ expandedRowRender, defaultExpandedRowKeys: ['0'] }}
       dataSource={data}
     />
   );
